Add script to set metadata
